-- Now with the option for using multiple buffers when we have multiple instances. -- Changes timestampes to uin32_t from uint64_t -- updated tests

+++ src/receivers/MAVLinkDefaultDeactivateRX.h
@@ -10,6 +10,6 @@ class MAVLinkDefaultDeactivateRX : public MAVLinkRXCallback
 public:
     MAVLinkDefaultDeactivateRX(MAVLinkEndpoint *e) : MAVLinkRXCallback(0, "default", e) {};
     bool process(mavlink_message_t *msg, mavlink_status_t *stat);
-    void timeout(uint64_t t) {};
-    bool isTimeout(uint64_t t) {return false;};
+    void timeout(uint32_t t) {};
+    bool isTimeout(uint32_t t) {return false;};
 };
