+++ src/receivers/MAVLinkHeartbeatRX.h
@@ -11,6 +11,6 @@ class MAVLinkHeartbeatRX : public MAVLinkRXCallback
 public:
     MAVLinkHeartbeatRX (MAVLinkEndpoint *e) : MAVLinkRXCallback(0, "heartbeat", e) {};
     bool process(mavlink_message_t *msg, mavlink_status_t *stat);
-    void timeout(uint64_t t) {};
-    bool isTimeout(uint64_t t) {return false;};
+    void timeout(uint32_t t) {};
+    bool isTimeout(uint32_t t) {return false;};
 };
